@@ -93,8 +93,8 @@ tasks.setupDevWorkspace.dependsOn copyMods
 tasks.setupDecompWorkspace.dependsOn copyMods
 
 
-sourceCompatibility = 1.7
-targetCompatibility = 1.7
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 
 task sourceJar(type: Jar) {
